+++ src/main/java/org/ezcode/codetest/application/problem/dto/request/ProblemUpdateRequest.java
@@ -26,7 +26,7 @@ public record ProblemUpdateRequest(
 	Long memoryLimit,
 
 	@Schema(description = "시간 제한", example = "1000.0")
-	Double timeLimit,
+	Long timeLimit,
 
 	@Schema(description = "출처", example = "ORIGINAL")
 	Reference reference
